begin porting to simpler fn
